Move UTF8 decoder logic into a separate file. Don't polyfill globally.
