+++ src/Microsoft.AspNetCore.Blazor.Browser.JS/src/Rendering/RenderBatch/OutOfProcessRenderBatch.ts
@@ -1,67 +1,5 @@
 import { RenderBatch, ArrayRange, RenderTreeDiff, ArrayValues, RenderTreeEdit, EditType, FrameType, RenderTreeFrame, RenderTreeDiffReader, RenderTreeFrameReader, RenderTreeEditReader, ArrayRangeReader, ArraySegmentReader, ArraySegment } from './RenderBatch';
-
-(function (scope) {
-  'use strict';
-
-  // bail early
-  if (scope['TextDecoder']) {
-    return false;
-  }
-
-  function TextDecoder(utfLabel = 'utf-8') {
-    if (utfLabel !== 'utf-8') {
-      throw new RangeError(
-        `Failed to construct 'TextDecoder': The encoding label provided ('${utfLabel}') is invalid.`);
-    }
-  }
-
-  TextDecoder.prototype.decode = function decode(bytes: Uint8Array) {
-    let pos = 0;
-    const len = bytes.length;
-    const out: number[] = [];
-
-    while (pos < len) {
-      const byte1 = bytes[pos++];
-      if (byte1 === 0) {
-        break;  // NULL
-      }
-
-      if ((byte1 & 0x80) === 0) {  // 1-byte
-        out.push(byte1);
-      } else if ((byte1 & 0xe0) === 0xc0) {  // 2-byte
-        const byte2 = bytes[pos++] & 0x3f;
-        out.push(((byte1 & 0x1f) << 6) | byte2);
-      } else if ((byte1 & 0xf0) === 0xe0) {
-        const byte2 = bytes[pos++] & 0x3f;
-        const byte3 = bytes[pos++] & 0x3f;
-        out.push(((byte1 & 0x1f) << 12) | (byte2 << 6) | byte3);
-      } else if ((byte1 & 0xf8) === 0xf0) {
-        const byte2 = bytes[pos++] & 0x3f;
-        const byte3 = bytes[pos++] & 0x3f;
-        const byte4 = bytes[pos++] & 0x3f;
-
-        // this can be > 0xffff, so possibly generate surrogates
-        let codepoint = ((byte1 & 0x07) << 0x12) | (byte2 << 0x0c) | (byte3 << 0x06) | byte4;
-        if (codepoint > 0xffff) {
-          // codepoint &= ~0x10000;
-          codepoint -= 0x10000;
-          out.push((codepoint >>> 10) & 0x3ff | 0xd800)
-          codepoint = 0xdc00 | codepoint & 0x3ff;
-        }
-        out.push(codepoint);
-      } else {
-        // FIXME: we're ignoring this
-      }
-    }
-
-    return String.fromCharCode.apply(null, out);
-  }
-
-  scope['TextDecoder'] = TextDecoder;
-
-}(window));
-
-const utf8Decoder = new TextDecoder('utf-8');
+import { decodeUtf8 } from './Utf8Decoder';
 
 const updatedComponentsEntryLength = 4; // Each is a single int32 giving the location of the data
 const referenceFramesEntryLength = 16; // 1 byte for frame type, then 3 bytes for type-specific data
@@ -234,7 +172,7 @@ class OutOfProcessStringReader {
         this.batchDataUint8.byteOffset + charsStart,
         numUtf8Bytes
       );
-      return utf8Decoder.decode(utf8Data);
+      return decodeUtf8(utf8Data);
     }
   }
 }
